+++ Script/DemoScriptText.swift
@@ -6,11 +6,14 @@ var demoScriptText = """
 # Select the section you want to run and then choose "Script ... Execute"
 #
 # You can run the whole file at once if you don't select a section, but since each
-# demo clears the log as it starts, the result is just like if you run the last
+# demo clears the log as it starts, the result is just like if you run the final
 # section alone.
 #
 # The demo race data is randomly generated, so you will see different results
 # as you run multiple times.
+#
+# NOTE: these examples all create random data using the script command `demo`.  To load a
+# real log file, use the menu command "File ... Load Log" or the script command `<`.
 
 # Example 1: create demo log and hilight race starts in Blue and crashes in Red
 demo