+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Pending Release
+- Fix script output view so that full available width is used.
+
 ## [1.4.1] 2024-07-14
 - Change File Load/Save command text so it's clear they refer to the script
 - Add new `File ... Load Log` menu command, to allow file loading when `<` won't work because full disk access is not enabled.